requirements.txt
pip
distlib
docker-py
click
